refactor: implement matching sun spinner, delete loginspinner
Made playlist-pages use sun spinner that matches other pages, and deleted old unused login spinner.
