--- frontend/src/components/login/Spinner.tsx
@@ -1,7 +1,0 @@
-export default function Spinner() {
-  return (
-    <div className="flex justify-center items-center mt-20">
-      <div className="w-10 h-10 border-4 border-gray-300 border-t-blue-500 rounded-full animate-spin"></div>
-    </div>
-  );
-}
